Move InstallRequirement.from_editable to a constructors module
