+++ tests/unit/test_req_file.py
@@ -12,6 +12,7 @@ from pip._internal.exceptions import (
     InstallationError, RequirementsFileParseError,
 )
 from pip._internal.index import PackageFinder
+from pip._internal.req.constructors import install_req_from_editable
 from pip._internal.req.req_file import (
     break_args_options, ignore_comments, join_lines, parse_requirements,
     preprocess, process_line, skip_regex,
@@ -218,7 +219,7 @@ class TestProcessLine(object):
         line = '-e %s' % url
         filename = 'filename'
         comes_from = '-r %s (line %s)' % (filename, 1)
-        req = InstallRequirement.from_editable(url, comes_from=comes_from)
+        req = install_req_from_editable(url, comes_from=comes_from)
         assert repr(list(process_line(line, filename, 1))[0]) == repr(req)
 
     def test_yield_editable_constraint(self):
@@ -226,7 +227,7 @@ class TestProcessLine(object):
         line = '-e %s' % url
         filename = 'filename'
         comes_from = '-c %s (line %s)' % (filename, 1)
-        req = InstallRequirement.from_editable(
+        req = install_req_from_editable(
             url, comes_from=comes_from, constraint=True)
         found_req = list(process_line(line, filename, 1, constraint=True))[0]
         assert repr(found_req) == repr(req)
